Move classes around according to updated directory/namespace structure.
